RF: Move make_bids_basename to utils.py
