+++ mne_bids/read.py
@@ -28,7 +28,8 @@ from mne_bids.utils import (_parse_bids_filename, _extract_landmarks,
                             _find_matching_sidecar, _parse_ext,
                             _get_ch_type_mapping, make_bids_folders,
                             _estimate_line_freq, _scale_coord_to_meters,
-                            _get_kinds_for_sub)
+                            _get_kinds_for_sub, make_bids_basename)
+
 
 reader = {'.con': io.read_raw_kit, '.sqd': io.read_raw_kit,
           '.fif': io.read_raw_fif, '.pdf': io.read_raw_bti,
@@ -686,8 +687,6 @@ def get_matched_empty_room(bids_basename, bids_root):
     if best_er_fname is None:
         er_basename = None
     else:
-        from mne_bids import make_bids_basename  # Avoid circular import.
-
         params = _parse_bids_filename(best_er_fname, verbose='warning')
         er_basename = make_bids_basename(
             subject=params.get('sub', None),
