+++ mne_bids/tests/test_copyfiles.py
@@ -19,6 +19,7 @@ with warnings.catch_warnings():
 
 from mne.datasets import testing
 from mne.utils import _TempDir
+from mne_bids import make_bids_basename
 from mne_bids.utils import _handle_kind, _parse_ext
 
 from mne_bids.copyfiles import (_get_brainvision_encoding,
@@ -27,7 +28,6 @@ from mne_bids.copyfiles import (_get_brainvision_encoding,
                                 copyfile_eeglab,
                                 copyfile_kit)
 
-from mne_bids.write import make_bids_basename
 
 base_path = op.join(op.dirname(mne.__file__), 'io')
